max room

+++ desafio_quality/src/main/java/desafio_quality/desafio_quality/dto/request/NeighborhoodRequestDTO.java
@@ -5,11 +5,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 public class NeighborhoodRequestDTO {
     private String name;
-    private Double priceByArea;
+    private BigDecimal priceByArea;
 }
